// frontend/src/types/generated/LabelPreviewRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BackLabelContent } from "./BackLabelContent";
import type { FrontLabelContent } from "./FrontLabelContent";

export type LabelPreviewRequest = { size: string, front: FrontLabelContent | null, back: BackLabelContent | null, };